+++ vf_interaction/config/vehicles.lua
@@ -3,6 +3,8 @@ DefaultVehicles = {
         "BALLER",
         "BANSHEE",
         "BATI",
+        "BATI2",
+        "casco",
         "BESTIAGTS",
         "FLASHGT",
         "FUROREGT",
@@ -30,4 +32,8 @@ DefaultVehicles = {
         "phantom2",
         "technical",
         "technical2",
+        "thruster",
+        "viseris",
+        "nero",
+        "terbyte"
     }