+++ vf_mtest/missions/cl_dispatch.lua
@@ -15,6 +15,8 @@ local SpawnedEnemies = false
 DecorRegister("m_dispatch_entity", 2)
 
 local enemyGroup = AddRelationshipGroup("m_dispatch_enemygroup")
+SetRelationshipBetweenGroups(5, GetHashKey("m_dispatch_enemygroup"), GetHashKey("PLAYER"))
+SetRelationshipBetweenGroups(5, GetHashKey("PLAYER"), GetHashKey("m_dispatch_enemygroup"))
 
 local function CreateTargetVehicle(model, x, y, z, rot)
     local modelHash = tostring(model)
@@ -124,10 +126,10 @@ function MissionDispatch.Init()
         SetPedIntoVehicle(playerPed, missionVeh, -1)
         SetVehicleHasBeenOwnedByPlayer(missionVeh, true)
         SetMaxWantedLevel(0)
-
+        SetVehicleRadioEnabled(playerPed, false)
         SetModelAsNoLongerNeeded(vehicleHash)
 
-       -- TriggerMusicEvent("MP_MC_CMH_SILO_PREP_START")
+        TriggerMusicEvent("MP_MC_CMH_SILO_PREP_START")
         TriggerMusicEvent("MP_MC_CMH_ACTION")
         missionStage = 1
     end
@@ -180,6 +182,7 @@ function MissionDispatch.Tick()
                 BeginTextCommandPrint("FM_IHELP_LCP")
                 EndTextCommandPrint(0.1, true)
             else
+                payOut = true
                 Missions.Kill()
             end
         end
@@ -199,6 +202,11 @@ function MissionDispatch.Kill()
 
     if DoesEntityExist(vehicle) then
         DeleteVehicle(vehicle)
+    end
+
+    if payOut then
+        TriggerServerEvent('vf_mtest:playercut', GetRandomIntInRange(5000, 20000))
+        payOut = false
     end
     
     ClearPrints()