+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog    
 
-## 1.0.0
+## 1.1.1
+
+- Fixed death screen loop when the player changed his skin.
+- Removed NativeUI from vf_base.
+
+## 1.1
+
+- Fixed random bugs.
+- Players can now use vending machines to restore health when `vf_vending` is loaded.
+- Players can use ATM's to deposit and withdraw money.
+- Dropped the interaction menu from the base resource and created a separated resource `vf_interaction`.
+
+## 1.0
 
 - Added Native phone using scaleforms instead of NUI.
 - Added multiple missions.
@@ -9,3 +21,4 @@
 - Added idle timer.
 - Added Ammu-Nation stores.
 - Added movie theaters.
+- Added loading screen